+++ src/device-types/HASensor.h
@@ -66,9 +66,10 @@ public:
 
     /**
      * Sets class of the state for the long term stats.
+     * You can find list of available values here: https://developers.home-assistant.io/docs/core/entity/sensor/#available-state-classes
      * See: https://developers.home-assistant.io/docs/core/entity/sensor/#long-term-statistics
      *
-     * @param stateClass The class name.
+     * @param stateClass The state class name.
      */
     inline void setStateClass(const char* stateClass)
         { _stateClass = stateClass; }